feat: add product image

+++ app/Services/ProductService.php
@@ -1,15 +1,19 @@
 <?php
+
 namespace App\Services;
 
 use App\Repositories\Interfaces\ProductRepositoryInterface;
+use Illuminate\Support\Arr;
 
 class ProductService
 {
     private $productRepo;
+    private $imageService;
 
-    public function __construct(ProductRepositoryInterface $productRepo)
+    public function __construct(ProductRepositoryInterface $productRepo, ImageService $imageService)
     {
         $this->productRepo = $productRepo;
+        $this->imageService = $imageService;
     }
 
     public function getAllProduct()
@@ -24,16 +28,31 @@ class ProductService
 
     public function saveProduct($request)
     {
-        $this->productRepo->createProduct($request);
+        $image = null;
+        if (Arr::exists($request, 'image')) {
+            $image = $this->imageService->saveImage($request['image']);
+        }
+        $this->productRepo->createProduct($request, $image);
     }
 
     public function updateProduct($request, $id)
     {
-        $this->productRepo->updateProduct($request, $id);
+        $productDetail = $this->productRepo->getProductById($id);
+        $image = $productDetail->image;
+        if (Arr::exists($request, 'image')) {
+            $image = $this->imageService->saveImage($request['image']);
+            if ($productDetail->image != null) {
+                $this->imageService->deleteImage($productDetail->image);
+            }
+        }
+        $this->productRepo->updateProduct($request, $id, $image);
     }
 
     public function deleteProduct($id)
     {
-        $this->productRepo->deleteProduct($id);
+        $productDetail = $this->productRepo->getProductById($id);
+        if ($this->imageService->deleteImage($productDetail->image)) {
+            $this->productRepo->deleteProduct($id);
+        }
     }
 }
